Added multi processing

+++ service/build.py
@@ -6,63 +6,60 @@ class BuildSync(object):
         self.dest_rack   = dest_rack
         self.logger      = logger
 
-    def sync(self, source_apps=None):
-        source_apps = source_apps if source_apps else self.source_rack.apps.get()
+    def sync(self, app=None):
+        self._sync(
+            self._compare(app)
+        )
 
-        requiring_update = self._compare(source_apps)
-
-        self._sync(requiring_update)
-
-    def _sync(self, apps):
+    def _sync(self, app):
         """ Export build artifact from source app and import it to the destination app """
+        if not app:
+            return
 
-        for app in apps:
-            app_name = app['name']
+        app_name = app['name']
 
-            self.logger.info('Syncing build for {}.{}'.format(self.dest_rack.name(), app_name))
+        self.logger.info('Syncing build for {}.{}'.format(self.dest_rack.name(), app_name))
 
-            active_build_id = self.source_rack.app(app_name).builds.active_build_id()
-            tmp_file        = '/tmp/{}.tgz'.format(active_build_id)
+        active_build_id = self.source_rack.app(app_name).builds.active_build_id()
+        tmp_file        = '/tmp/{}.tgz'.format(active_build_id)
 
-            self.source_rack.app(app_name).builds.export_build(active_build_id, tmp_file)
-            self.dest_rack.app(app_name).builds.import_build(active_build_id, tmp_file)
-            os.remove(tmp_file)
+        self.source_rack.app(app_name).builds.export_build(active_build_id, tmp_file)
+        self.dest_rack.app(app_name).builds.import_build(active_build_id, tmp_file)
+        os.remove(tmp_file)
 
-        return apps
+        return app
 
-    def _compare(self, apps):
+    def _compare(self, app):
         """ Compare source and destination apps to see if the build id are the same or not """
 
-        self.logger.info('Comparing builds')
+        if not app:
+            return
 
-        mismatched_build = []
+        app_name = app['name']
 
-        for app in apps:
-            app_name = app['name']
+        self.logger.info('Comparing build for app {}'.format(app_name))
 
-            source_build_id = self.source_rack.app(app_name).builds.active_build_id()
+        source_build_id = self.source_rack.app(app_name).builds.active_build_id()
 
-            if not source_build_id:
-                continue
+        if not source_build_id:
+            return None
 
-            #check if this build already exists on the dest rack
-            dest_build = self.dest_rack.app(app_name).builds.get(source_build_id)
+        #check if this build already exists on the dest rack
+        dest_build = self.dest_rack.app(app_name).builds.get(source_build_id)
 
-            if dest_build and 'error' not in dest_build:
-                self.logger.info('Build {} exists on {}.{}'.format(
-                    source_build_id,
-                    self.dest_rack.name(),
-                    app_name
-                ))
-
-                continue
-
-            self.logger.info('Build {} not found on {}.{}'.format(
+        if dest_build and 'error' not in dest_build:
+            self.logger.info('Build {} exists on {}.{}'.format(
                 source_build_id,
                 self.dest_rack.name(),
                 app_name
             ))
 
-            mismatched_build.append(app)
+            return None
 
-        return mismatched_build
+        self.logger.info('Build {} not found on {}.{}'.format(
+            source_build_id,
+            self.dest_rack.name(),
+            app_name
+        ))
+
+        return app
